Fix calculating frequency for note

+++ src/main/kotlin/it/krzeminski/fsynth/MusicNote.kt
@@ -94,7 +94,8 @@ enum class MusicNote(private val midiNoteNumber: Int) {
 
     val frequency: Float
         get() =
-            2.0f.pow((midiNoteNumber - CONCERT_A_NOTE_MIDI_NOTE_NUMBER)/SEMITONES_IN_OCTAVE)*CONCERT_A_NOTE_FREQUENCY
+            2.0f.pow((midiNoteNumber - CONCERT_A_NOTE_MIDI_NOTE_NUMBER).toFloat()/SEMITONES_IN_OCTAVE)*
+                    CONCERT_A_NOTE_FREQUENCY
 
     companion object {
         const val CONCERT_A_NOTE_MIDI_NOTE_NUMBER = 69
